New changes, who would have thought

+++ src/TestHub.ApplicationCore/Entities/AnswerForm.cs
@@ -4,6 +4,6 @@ namespace TestHub.ApplicationCore.Entities
 {
     public abstract class AnswerForm: BaseEntity
     {
-        public abstract int Grade();
+        public abstract decimal Grade();
     }
 }
